requirements.txt
langchain
langchain-groq
langchain-community
tavily-python
faiss-cpu
python-dotenv
rich
beautifulsoup4